qa/parity-test/requirements.txt
pytest>=8.0
pytest-json-report>=1.5
requests>=2.31
pymysql>=1.1
psycopg2-binary>=2.9
deepdiff>=7.0
rich>=13.0
